+++ HitchFix/Models/Problem.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string ProblemName { get; set; }
-        public int Price { get; set; }
-        public int? DiscountPrice { get; set; }
-        public int TotalPriceAfterDiscount { get; set; }
+        public double Price { get; set; }
+        public double? DiscountPrice { get; set; }
+        public double TotalPriceAfterDiscount { get; set; }
     }
 }
